requirements.txt
requests == 2.32.3
beautifulsoup4 == 4.13.1
pydantic == 2.10.6
python-dotenv == 1.0.1
